requirements.txt
flask>=3.0.0
openai>=1.0.0
pillow>=10.0.0
python-multipart>=0.0.6
werkzeug>=3.0.0
python-dotenv>=1.0.0
